refactor socket names
